move out fuse metrics
