+++ src/common/metrics/src/metrics/mod.rs
@@ -17,4 +17,5 @@ pub mod cluster;
 pub mod http;
 pub mod interpreter;
 pub mod openai;
+pub mod session;
 pub mod storage;
