+++ src/query/service/src/sessions/mod.rs
@@ -18,7 +18,6 @@ mod query_ctx_shared;
 mod session;
 mod session_ctx;
 mod session_info;
-mod session_metrics;
 mod session_mgr;
 mod session_mgr_status;
 mod session_privilege_mgr;
