+++ src/query/storages/common/cache/Cargo.toml
@@ -11,9 +11,6 @@ edition = { workspace = true }
 doctest = false
 test = false
 
-[features]
-enable-histogram-metrics = ["metrics/enable-histogram"]
-
 [dependencies]
 bytes = { workspace = true }
 common-cache = { path = "../../../../common/cache" }
@@ -25,11 +22,8 @@ async-trait = { version = "0.1.57", package = "async-trait-fn" }
 crc32fast = "1.3.2"
 crossbeam-channel = "0.5.6"
 hex = "0.4.3"
-lazy_static = { workspace = true }
 log = { workspace = true }
-metrics = "0.20.1"
 parking_lot = "0.12.1"
-prometheus-client = { workspace = true }
 siphasher = "0.3.10"
 
 [dev-dependencies]
