+++ src/query/storages/fuse/src/io/read/block/block_reader_native_deserialize.rs
@@ -37,6 +37,7 @@ use common_exception::ErrorCode;
 use common_exception::Result;
 use common_expression::ColumnId;
 use common_expression::DataBlock;
+use common_metrics::storage::*;
 use common_storage::ColumnNode;
 use storages_common_cache::CacheAccessor;
 use storages_common_cache::TableDataCacheKey;
@@ -50,7 +51,6 @@ use crate::io::read::block::block_reader_merge_io::DataItem;
 use crate::io::BlockReader;
 use crate::io::NativeReaderExt;
 use crate::io::UncompressedBuffer;
-use crate::metrics::*;
 
 impl BlockReader {
     /// Deserialize column chunks data from native format to DataBlock.
