+++ src/query/storages/fuse/src/lib.rs
@@ -27,7 +27,6 @@ mod fuse_column;
 mod fuse_part;
 mod fuse_table;
 mod fuse_type;
-mod metrics;
 
 pub mod io;
 pub mod operations;
