+++ src/query/storages/fuse/src/operations/common/processors/sink_commit.rs
@@ -28,6 +28,7 @@ use common_exception::Result;
 use common_expression::BlockMetaInfoDowncast;
 use common_meta_app::schema::TableInfo;
 use common_meta_app::schema::UpsertTableCopiedFileReq;
+use common_metrics::storage::*;
 use common_pipeline_core::LockGuard;
 use log::debug;
 use log::error;
@@ -43,10 +44,6 @@ use storages_common_table_meta::meta::TableSnapshot;
 use storages_common_table_meta::meta::Versioned;
 
 use crate::io::TableMetaLocationGenerator;
-use crate::metrics::metrics_inc_commit_aborts;
-use crate::metrics::metrics_inc_commit_copied_files;
-use crate::metrics::metrics_inc_commit_milliseconds;
-use crate::metrics::metrics_inc_commit_mutation_success;
 use crate::operations::common::AbortOperation;
 use crate::operations::common::CommitMeta;
 use crate::operations::common::SnapshotGenerator;
