+++ src/query/storages/fuse/src/operations/common/processors/transform_serialize_block.rs
@@ -22,6 +22,7 @@ use common_exception::ErrorCode;
 use common_exception::Result;
 use common_expression::BlockMetaInfoDowncast;
 use common_expression::DataBlock;
+use common_metrics::storage::*;
 use common_pipeline_core::pipe::PipeItem;
 use common_pipeline_core::processors::port::InputPort;
 use common_pipeline_core::processors::processor::ProcessorPtr;
@@ -31,12 +32,6 @@ use storages_common_index::BloomIndex;
 use crate::io::write_data;
 use crate::io::BlockBuilder;
 use crate::io::BlockSerialization;
-use crate::metrics::metrics_inc_block_index_write_bytes;
-use crate::metrics::metrics_inc_block_index_write_milliseconds;
-use crate::metrics::metrics_inc_block_index_write_nums;
-use crate::metrics::metrics_inc_block_write_bytes;
-use crate::metrics::metrics_inc_block_write_milliseconds;
-use crate::metrics::metrics_inc_block_write_nums;
 use crate::operations::common::BlockMetaIndex;
 use crate::operations::common::MutationLogEntry;
 use crate::operations::common::MutationLogs;
