+++ src/query/storages/fuse/src/operations/delete.rs
@@ -42,15 +42,13 @@ use common_expression::TableSchema;
 use common_expression::Value;
 use common_expression::ROW_ID_COL_NAME;
 use common_functions::BUILTIN_FUNCTIONS;
+use common_metrics::storage::*;
 use common_sql::evaluator::BlockOperator;
 use storages_common_index::RangeIndex;
 use storages_common_pruner::RangePruner;
 use storages_common_table_meta::meta::StatisticsOfColumns;
 use storages_common_table_meta::meta::TableSnapshot;
 
-use crate::metrics::metrics_inc_deletion_block_range_pruned_nums;
-use crate::metrics::metrics_inc_deletion_block_range_pruned_whole_block_nums;
-use crate::metrics::metrics_inc_deletion_segment_range_purned_whole_segment_nums;
 use crate::operations::mutation::Mutation;
 use crate::operations::mutation::MutationAction;
 use crate::operations::mutation::MutationPartInfo;
