+++ src/query/storages/fuse/src/operations/mutation/recluster_aggregator.rs
@@ -25,6 +25,7 @@ use common_expression::BlockMetaInfoDowncast;
 use common_expression::BlockMetaInfoPtr;
 use common_expression::BlockThresholds;
 use common_expression::DataBlock;
+use common_metrics::storage::*;
 use common_pipeline_transforms::processors::transforms::AsyncAccumulatingTransform;
 use itertools::Itertools;
 use opendal::Operator;
@@ -36,7 +37,6 @@ use storages_common_table_meta::meta::Versioned;
 use crate::io::SegmentsIO;
 use crate::io::SerializedSegment;
 use crate::io::TableMetaLocationGenerator;
-use crate::metrics::metrics_inc_recluster_write_block_nums;
 use crate::operations::common::AbortOperation;
 use crate::operations::common::CommitMeta;
 use crate::operations::common::ConflictResolveContext;
