+++ src/query/storages/fuse/src/operations/recluster.rs
@@ -28,6 +28,7 @@ use common_expression::DataSchemaRefExt;
 use common_expression::SortColumnDescription;
 use common_expression::TableSchemaRef;
 use common_meta_app::schema::CatalogInfo;
+use common_metrics::storage::*;
 use common_pipeline_core::processors::processor::ProcessorPtr;
 use common_pipeline_transforms::processors::transforms::build_merge_sort_pipeline;
 use common_pipeline_transforms::processors::transforms::AsyncAccumulatingTransformer;
@@ -39,9 +40,6 @@ use log::warn;
 use opendal::Operator;
 use storages_common_table_meta::meta::CompactSegmentInfo;
 
-use crate::metrics::metrics_inc_recluster_block_bytes_to_read;
-use crate::metrics::metrics_inc_recluster_block_nums_to_read;
-use crate::metrics::metrics_inc_recluster_row_nums_to_read;
 use crate::operations::common::CommitSink;
 use crate::operations::common::MutationGenerator;
 use crate::operations::common::TransformSerializeBlock;
