+++ src/query/storages/fuse/src/operations/replace_into/mutator/merge_into_mutator.rs
@@ -32,6 +32,7 @@ use common_expression::DataBlock;
 use common_expression::FieldIndex;
 use common_expression::Scalar;
 use common_expression::TableSchema;
+use common_metrics::storage::*;
 use common_sql::evaluator::BlockOperator;
 use common_sql::executor::OnConflictField;
 use log::info;
@@ -55,22 +56,6 @@ use crate::io::CompactSegmentInfoReader;
 use crate::io::MetaReaders;
 use crate::io::ReadSettings;
 use crate::io::WriteSettings;
-use crate::metrics::metrics_inc_replace_accumulated_merge_action_time_ms;
-use crate::metrics::metrics_inc_replace_apply_deletion_time_ms;
-use crate::metrics::metrics_inc_replace_block_number_after_pruning;
-use crate::metrics::metrics_inc_replace_block_number_bloom_pruned;
-use crate::metrics::metrics_inc_replace_block_number_totally_loaded;
-use crate::metrics::metrics_inc_replace_block_number_write;
-use crate::metrics::metrics_inc_replace_block_of_zero_row_deleted;
-use crate::metrics::metrics_inc_replace_deleted_blocks_rows;
-use crate::metrics::metrics_inc_replace_number_accumulated_merge_action;
-use crate::metrics::metrics_inc_replace_number_apply_deletion;
-use crate::metrics::metrics_inc_replace_replaced_blocks_rows;
-use crate::metrics::metrics_inc_replace_row_number_after_pruning;
-use crate::metrics::metrics_inc_replace_row_number_totally_loaded;
-use crate::metrics::metrics_inc_replace_row_number_write;
-use crate::metrics::metrics_inc_replace_segment_number_after_pruning;
-use crate::metrics::metrics_inc_replace_whole_block_deletion;
 use crate::operations::acquire_task_permit;
 use crate::operations::common::BlockMetaIndex;
 use crate::operations::common::MutationLogEntry;
