+++ src/query/storages/fuse/src/operations/replace_into/processors/processor_unbranched_replace_into.rs
@@ -25,6 +25,7 @@ use common_expression::DataBlock;
 use common_expression::FieldIndex;
 use common_expression::RemoteExpr;
 use common_expression::TableSchema;
+use common_metrics::storage::*;
 use common_pipeline_core::pipe::Pipe;
 use common_pipeline_core::pipe::PipeItem;
 use common_pipeline_core::processors::port::InputPort;
@@ -35,7 +36,6 @@ use common_pipeline_core::processors::Processor;
 use common_sql::executor::OnConflictField;
 use storages_common_table_meta::meta::ColumnStatistics;
 
-use crate::metrics::metrics_inc_replace_process_input_block_time_ms;
 use crate::operations::replace_into::mutator::mutator_replace_into::ReplaceIntoMutator;
 
 pub struct UnbranchedReplaceIntoProcessor {
