+++ src/query/storages/fuse/src/pruning/segment_pruner.rs
@@ -16,13 +16,10 @@ use std::sync::Arc;
 
 use common_exception::Result;
 use common_expression::TableSchemaRef;
+use common_metrics::storage::*;
 use storages_common_table_meta::meta::CompactSegmentInfo;
 
 use crate::io::SegmentsIO;
-use crate::metrics::metrics_inc_bytes_segment_range_pruning_after;
-use crate::metrics::metrics_inc_bytes_segment_range_pruning_before;
-use crate::metrics::metrics_inc_segments_range_pruning_after;
-use crate::metrics::metrics_inc_segments_range_pruning_before;
 use crate::pruning::PruningContext;
 use crate::pruning::SegmentLocation;
