+++ src/query/storages/system/Cargo.toml
@@ -11,9 +11,6 @@ edition = { workspace = true }
 doctest = false
 test = false
 
-[features]
-enable-histogram-metrics = ["common-metrics/enable-histogram"]
-
 [dependencies]
 common-ast = { path = "../../ast" }
 common-base = { path = "../../../common/base" }
